Move over `crate_label_for_command`
