+++ crates/assistant_slash_commands/src/diagnostics_command.rs
@@ -21,7 +21,7 @@ use util::paths::PathMatcher;
 use util::ResultExt;
 use workspace::Workspace;
 
-use crate::slash_command::create_label_for_command;
+use crate::create_label_for_command;
 
 pub(crate) struct DiagnosticsSlashCommand;
